PatchUnicode-1046 v0.13 コンボボックスの履歴削除に対応

+++ sakura_core/dlg/CDialog.cpp
@@ -762,6 +762,10 @@ static void DeleteRecentItem(
 
 		// 履歴項目を削除
 		int nRecentIndex = pRecent->FindItemByText( cItemText.GetStringPtr() );
+		// お気に入りチェック
+		if( pRecent->IsFavorite(nRecentIndex) ){
+			return;
+		}
 		if( 0 <= nRecentIndex ){
 			pRecent->DeleteItem(nRecentIndex);
 		}
